app/requirements.txt
pandas==2.0.1
numpy==1.26.4
python-multipart
lz4==4.4.4
packaging
fastapi==0.110.0
uvicorn==0.29.0
joblib==1.3.2
faiss-cpu==1.8.0  # Critical for ARM compatibility